Removed Hardcore incompatibilities. Removed Voracious Desire. Optimizations.
Removed all overrides to Ingestible items and the HC_Manager script, so we are now compatible with mods that must change those.

Also gotten rid of the Voracious Desire addiction. We may want to implement a Cannibal toggle at some point to make it so cannibalism does not carry an addiction, to make everything line up conceptually.

Deleted the Hardcore folder entirely. VoreSurvivalScript now handles health regen and Survival satiation, is now in \User. We should use Namespaces in future.

Various property optimisation to scripts.